address comments

+++ src/ray/core_worker/task_interface.cc
@@ -93,7 +93,7 @@ void ActorHandle::ClearNewActorHandles() { new_actor_handles_.clear(); }
 CoreWorkerTaskInterface::CoreWorkerTaskInterface(
     WorkerContext &worker_context, std::unique_ptr<RayletClient> &raylet_client)
     : worker_context_(worker_context) {
-  task_submitters_.emplace(static_cast<int>(TaskTransportType::RAYLET),
+  task_submitters_.emplace(TaskTransportType::RAYLET,
                            std::unique_ptr<CoreWorkerRayletTaskSubmitter>(
                                new CoreWorkerRayletTaskSubmitter(raylet_client)));
 }
@@ -134,7 +134,7 @@ Status CoreWorkerTaskInterface::SubmitTask(const RayFunction &function,
                                            std::vector<ObjectID> *return_ids) {
   auto builder = BuildCommonTaskSpec(function, args, task_options.num_returns,
                                      task_options.resources, {}, return_ids);
-  return task_submitters_[static_cast<int>(TaskTransportType::RAYLET)]->SubmitTask(
+  return task_submitters_[TaskTransportType::RAYLET]->SubmitTask(
       builder.Build());
 }
 
@@ -155,7 +155,7 @@ Status CoreWorkerTaskInterface::CreateActor(
   (*actor_handle)->IncreaseTaskCounter();
   (*actor_handle)->SetActorCursor(return_ids[0]);
 
-  return task_submitters_[static_cast<int>(TaskTransportType::RAYLET)]->SubmitTask(
+  return task_submitters_[TaskTransportType::RAYLET]->SubmitTask(
       builder.Build());
 }
 
@@ -188,7 +188,7 @@ Status CoreWorkerTaskInterface::SubmitActorTask(ActorHandle &actor_handle,
   guard.unlock();
 
   // Submit task.
-  auto status = task_submitters_[static_cast<int>(TaskTransportType::RAYLET)]->SubmitTask(
+  auto status = task_submitters_[TaskTransportType::RAYLET]->SubmitTask(
       builder.Build());
 
   // Remove cursor from return ids.
